p4i integration Integrate 260435 from 4.3 to main: Fixes:    Compile when qreal is float RevBy:    Jesper
[git-p4: depot-paths = "//depot/qt/main/": change = 260436]

+++ tests/auto/qtessellator/oldtessellator.cpp
@@ -198,8 +198,8 @@ void old_tesselate_polygon(QVector<XTrapezoid> *traps, const QPointF *pg, int pg
 	edge.b = p1.y() - edge.m * p1.x(); // intersection with y axis
 	edge.m = edge.m != 0.0 ? 1.0 / edge.m : 0.0; // inverted derivative
 	edges.append(edge);
-        ymin = qMin(ymin, XFixedToDouble(edge.p1.y));
-        ymax = qMax(ymax, XFixedToDouble(edge.p2.y));
+        ymin = qMin(ymin, qreal(XFixedToDouble(edge.p1.y)));
+        ymax = qMax(ymax, qreal(XFixedToDouble(edge.p2.y)));
     }
 
     QList<const QEdge *> et; 	    // edge list
